moving functions out to seperate files
